Fix ToDo list storage error for multiple lists on a single page
Currently, the update method uses the last global value for form, hence works only for the last todo list on page (for todo items from all other lists, a state of "unchecked" will be assumed and updated in local storage. Loading is not affected and works correctly for multiple lists.
My change massively simplifies the update method and removes the need to know the form a todo item belongs to. Instead, it directly accesses the checked attribute of the event source and uses it to update local storage.